fix: add newline at the end of file
Most editors are configured to add a newline at the end of the file while Babel strips the last line when transforming. This means that if we manually edit the output fixtures, there's always a mismatch in the code due to the extra new line. We workaround this by always adding a new line at the end of the file